Added context on react components
