+++ Todo.md
@@ -5,6 +5,8 @@
 - [ ] Rebuild all front end structure with "React context"
 - [ ] Do a better and scalable component division
 - [ ] Make all style responsive
+- [ ] Fix date refresh in events
+- [ ] Refresh calendar when changes are made
 
 ---
