+++ front-end/src/pages/home.js
@@ -2,6 +2,11 @@ import Header from '../components/Header';
 import Hero from '../components/Hero';
 
 const Home = () => {
+  /* Create a state here to share user data
+   * - If user is authenticated or not
+   * - al user util data
+   */
+
   return (
     <div className="home">
       <Header />
